Fix game detail client layout and devide into each vue.js component
