+++ client/src/components/GameInfo.vue
@@ -1,66 +1,24 @@
 <template>
   <v-card>
     <!-- Game title and description -->
+    <v-card-media :src="game.img" height="340px">
+    </v-card-media>
     <v-card-title primary-title>
       <div>
-        <h4 class="display-2">{{ game.title }}</h4>
-        <div class="ml-2">{{ game.description }}</div>
+        <h3 class="headline mb-0">{{ game.title }}</h3>
+        <div>{{ game.description }}</div>
       </div>
     </v-card-title>
 
     <!-- Game instructions -->
     <v-card-text>
       <div>
-        <h5 class="title">Instructions</h5>
-        <ul>
+        <h5 class="title">遊び方</h5>
+        <ol>
           <li v-for="(is, index) in game.instructions" :key="index">
-            {{ is.inst }}
+            {{ is.explanation }}
           </li>
-        </ul>
-      </div>
-    </v-card-text>
-
-    <!-- Access count -->
-    <v-card-text>
-      <div>
-        <h5 class="title">総アクセス数</h5>
-        <div class="ml-2">{{ game.access }} 回</div>
-      </div>
-    </v-card-text>
-
-    <!-- Stars count -->
-    <v-card-text>
-      <div>
-        <h5 class="title">総合評価</h5>
-        <div class="ml-2">
-          <v-icon v-for="n in game.star" :key="n" class="orange--text">star</v-icon>
-        </div>
-      </div>
-    </v-card-text>
-
-    <!-- User commets -->
-    <v-card-text>
-      <div>
-        <h5 class="title">最新口コミ</h5>
-
-       <v-expansion-panel expand>
-          <v-expansion-panel-content v-for="(rv, index) in game.reviews" :key="index">
-            <div slot="header">
-              {{ rv.user }}
-              <span>
-                <v-icon v-for="n in rv.star" :key="n" class="yellow--text star">
-                  star
-                </v-icon>
-              </span>
-            </div>
-            <v-card>
-              <v-card-text class="grey lighten-3">
-                {{ rv.comment }}
-              </v-card-text>
-            </v-card>
-          </v-expansion-panel-content>
-        </v-expansion-panel>
-
+        </ol>
       </div>
     </v-card-text>
   </v-card>
@@ -68,7 +26,7 @@
 
 <script>
 /**
- * Game Detail Information Component
+ * Game Information Component
  */
 export default {
   props: {
@@ -80,7 +38,4 @@ export default {
 </script>
 
 <style>
-.star {
-  font-size: 75%;
-}
 </style>
